Combine NumberUtils into DateUtils
