--- lib/emendate/number_utils.rb
@@ -1,46 +1,0 @@
-# frozen_string_literal: true
-
-module Emendate
-  module NumberUtils
-    extend self
-
-    def min_season
-      return 21 unless Emendate.options.max_month_number_handling == :months
-    end
-
-    def max_season
-      return 24 if Emendate.options.max_month_number_handling == :edtf_level_1
-      return 41 if Emendate.options.max_month_number_handling == :edtf_level_2
-    end
-
-    # NOTE: can be valid day but not valid in the context of a given date (Feb 29 in a non-leap year)
-    def valid_day?(str)
-      int = str.to_i
-      str.length <= 2 && int >= 1 && int <= 31
-    end
-
-    def valid_month?(int)
-      true if int > 0 && int < 13
-    end
-
-    def valid_month_or_season?(str)
-      (valid_month?(str) || valid_season?(str))
-    end
-
-    # EDTF season/quarter/semester indicators
-    def valid_season?(str)
-      return false if Emendate.options.max_month_number_handling == :months
-
-      int = str.to_i
-      return false if int < min_season
-      return false if int > max_season
-
-      true
-    end
-
-    def valid_year?(str)
-      str = str.to_s unless str.is_a?(String)
-      str.length <= 4
-    end
-  end
-end
